requirements.txt
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
mysql-connector-python==9.2.0
outcome==1.3.0.post0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.1.0
pycparser==2.22
pypdfium2==4.30.1
PySocks==1.7.1
requests==2.32.3
selenium==4.30.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.0
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
validators
ftfy
